**Description**: Closes https://github.com/buger/jsonparser/issues/117. You can now use `Set` with array indices when the full path does not already exist, and supports `[+]` for append and `[-]` for prepend.

**Updated Logic:**
- if the array and numeric index exists, update it
- if the array exists and [+], append it
- if the array exists and [-], prepend it
- if the array exists and numeric index does not, pad with `null` as needed
- if the array does not exist and [0] or [+/-], create the array and set at 0
- if the array does not exist and non-zero numeric index, pad with `null` as needed

**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                         50000            122002 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                       500000             23302 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserDeleteMedium-4                 500000             19750 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4         1000000             16974 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4          300000             22338 ns/op             672 B/op         12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4       200000             31673 ns/op             624 B/op         11 allocs/op
BenchmarkJsonParserSmall-4                       3000000              2227 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4          3000000              2255 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4          2000000              3258 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4       3000000              2926 ns/op             240 B/op          8 allocs/op
BenchmarkJsonParserSetSmall-4                    2000000              4072 ns/op             816 B/op          5 allocs/op
BenchmarkJsonParserDelSmall-4                    3000000              3447 ns/op               0 B/op          0 allocs/op
```

**Benchmark after change**:
```
BenchmarkJsonParserLarge-4                         50000            111986 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                       500000             21889 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserDeleteMedium-4                1000000             18926 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4          300000             20039 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4          500000             22492 ns/op             672 B/op         12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4       200000             31212 ns/op             624 B/op         11 allocs/op
BenchmarkJsonParserSmall-4                       5000000              2035 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4          3000000              2150 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4          2000000              3209 ns/op             256 B/op          9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4       3000000              2894 ns/op             240 B/op          8 allocs/op
BenchmarkJsonParserSetSmall-4                    2000000              4085 ns/op             816 B/op          5 allocs/op
BenchmarkJsonParserDelSmall-4                    2000000              3633 ns/op               0 B/op          0 allocs/op
```