@JoshKCarroll 2017-09-13
See initial notes and discussion on https://github.com/qntfy/jsonparser/pull/1, where first commit code came from. Thanks @chilland!