@ryanleary 2017-09-13
The proposed logic in the description LGTM