@JoshKCarroll 2017-09-15
Note on review: currently `Set()` returns `KeyPathNotFound` error if the top level json is an array. Only top level objects are supported. Not planning to add top level array support with this PR. If we wanted to add it we would do it here:
```
			// We can't set a top-level key if data isn't an object
			if len(data) == 0 || data[firstToken] != '{' {
				return nil, KeyPathNotFoundError
			}
```
Let me know if we'd like to add this to scope in the near term and I'm happy to file an issue / add it.