@JoshKCarroll 2017-09-15
@ryanleary @chilland this is now complete and ready for review. Do we have ability to squash in github or do I need to squash manually? Also can we turn on https://travis-ci.org/qntfy/jsonparser ?