@ryanleary 2017-09-15
Travis enabled. We can squash and merge here.