@ryanleary 2017-09-16
I *think* this looks good. A handful of clarifying comments will go a long way toward its understandability and maintainability I think.