@JoshKCarroll 2017-09-17
@ryanleary does this sufficiently clarify? Also, any thought on versioning? Should we do any release before this PR? Feels a little weird to set this at 1.0.0.