@ryanleary 2017-09-17
Looks good to me. Thanks for adding the comments. 

No strong opinions on versioning really. Probably makes sense to divorce any version tags in this repo from the original one since they won't stay in sync anyway?