@JoshKCarroll 2017-09-17
They never did any versioned releases so it's good there. We've just been versioning by commit.